packages/arena-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "arena-core"
version = "0.1.0"
description = "Shared core library for AI Arena game services"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "websockets>=11.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]
